// src/types/model/tmp1.model.ts
export type Type1 = {
  id: number;
  collection_slug: string;
  created_at: string;
  description: string;
  image_original_url: string;
  image_thumbnail_url: string;
  image_url: string;
  last_sale_created_date: null;
  last_sale_price: number | string | null;
  last_sale_symbol: string;
  name: string;
  num_sales: number;
  permalink: string;
  token_id: string;
};